ADD: Configuration

+++ src/main/java/de/goldendeveloper/backup/discord/ExportImport.java
@@ -6,16 +6,27 @@ import org.json.JSONArray;
 import org.json.JSONObject;
 
 import java.io.*;
+import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
 public class ExportImport {
 
     public static String PERMISSIONS = "permissions";
-    public static String TYPE = "Type";
+    public static String TYPE = "type";
     public static String ROLES = "roles";
     public static String POSITION = "position";
     public static String USERLIMIT = "userlimit";
+    public static String NAME = "name";
+    public static String DESCRIPTION = "description";
+    public static String TOPIC = "topic";
+    public static String CATEGORY = "category";
+    public static String CHANNELS = "channels";
+    public static String AFKCHANNEL = "afkchannel";
+    public static String AFKTIMEOUT = "afktimeout";
+    public static String SYSTEMMESSAGECHANNEL = "systemmessagechannel";
+    public static String DEFAULTNOTIFICATIONLEVEL = "defaultnotificationlevel";
+    public static String NSFW = "nsfw";
 
 
     public static File Export(Guild guild) {
@@ -72,7 +83,7 @@ public class ExportImport {
     public static @NotNull JSONObject exportBackup(@NotNull Guild guild) {
         JSONObject category = new JSONObject();
         for (Category c : guild.getCategories()) {
-            category.put(c.getName(), getChannel(c));
+            category.put(c.getName(), getCategory(c));
         }
 
         JSONObject roles = new JSONObject();
@@ -81,10 +92,18 @@ public class ExportImport {
         }
 
         JSONObject json = new JSONObject();
-        json.put("name", guild.getName());
-        json.put("description", guild.getDescription());
-        json.put("category", category);
+        json.put(NAME, guild.getName());
+        json.put(DESCRIPTION, guild.getDescription());
+        json.put(CATEGORY, category);
         json.put(ROLES, roles);
+        json.put(AFKTIMEOUT, guild.getAfkTimeout());
+        if (guild.getAfkChannel() != null) {
+            json.put(AFKCHANNEL, guild.getAfkChannel().getName());
+        }
+        if (guild.getSystemChannel() != null) {
+            json.put(SYSTEMMESSAGECHANNEL, guild.getSystemChannel().getName());
+        }
+        json.put(DEFAULTNOTIFICATIONLEVEL, guild.getDefaultNotificationLevel());
         return json;
     }
 
@@ -93,6 +112,13 @@ public class ExportImport {
         role.put(PERMISSIONS, "comming soon");
         role.put(POSITION, r.getPosition());
         return role;
+    }
+
+    public static JSONObject getCategory(Category c) {
+        JSONObject category = new JSONObject();
+        category.put(CHANNELS, getChannel(c));
+        category.put(POSITION, c.getPosition());
+        return category;
     }
 
     public static @NotNull JSONObject getChannel(@NotNull Category c) {
@@ -104,37 +130,40 @@ public class ExportImport {
     }
 
     public static @NotNull JSONObject getGuildChannel(@NotNull Channel guildChannel) {
-        JSONObject willkommen = new JSONObject();
+        JSONObject channelJson = new JSONObject();
         if (guildChannel.getType().equals(ChannelType.TEXT)) {
-            willkommen.put(TYPE, "text");
             TextChannel textChannel = guildChannel.getJDA().getTextChannelById(guildChannel.getId());
             if (textChannel != null) {
-                willkommen.put("Description", textChannel.getTopic());
-                willkommen.put(POSITION, textChannel.getPosition());
+                channelJson.put(TYPE, "text");
+                channelJson.put(NSFW, "" + textChannel.isNSFW() + "");
+                channelJson.put(POSITION, textChannel.getPosition());
+                channelJson.put(DESCRIPTION, textChannel.getTopic());
             }
         } else if (guildChannel.getType().equals(ChannelType.VOICE)) {
-            willkommen.put(TYPE, "voice");
             VoiceChannel voiceChannel = guildChannel.getJDA().getVoiceChannelById(guildChannel.getId());
             if (voiceChannel != null) {
-                willkommen.put(POSITION, voiceChannel.getPosition());
-                willkommen.put(USERLIMIT, voiceChannel.getUserLimit());
+                channelJson.put(TYPE, "voice");
+                channelJson.put(POSITION, voiceChannel.getPosition());
+                channelJson.put(USERLIMIT, voiceChannel.getUserLimit());
             }
         } else if (guildChannel.getType().equals(ChannelType.STAGE)) {
-            willkommen.put(TYPE, "stage");
             StageChannel stageChannel = guildChannel.getJDA().getStageChannelById(guildChannel.getId());
             if (stageChannel != null) {
-                willkommen.put(POSITION, stageChannel.getPosition());
+                channelJson.put(TYPE, "stage");
+                channelJson.put(POSITION, stageChannel.getPosition());
             }
         } else if (guildChannel.getType().equals(ChannelType.NEWS)) {
-            willkommen.put(TYPE, "news");
             NewsChannel newsChannel = guildChannel.getJDA().getNewsChannelById(guildChannel.getId());
             if (newsChannel != null) {
-                willkommen.put(POSITION, newsChannel.getPosition());
+                channelJson.put(TYPE, "news");
+                channelJson.put(POSITION, newsChannel.getPosition());
+                channelJson.put(NSFW, "" + newsChannel.isNSFW() + "");
+                channelJson.put(TOPIC, newsChannel.getTopic());
             }
         }
-        willkommen.put(ROLES, "Hier kommen die Rollen hin");
-        willkommen.put(PERMISSIONS, "comming soon");
-        return willkommen;
+        channelJson.put(ROLES, "Hier kommen die Rollen hin");
+        channelJson.put(PERMISSIONS, "comming soon");
+        return channelJson;
     }
 
     public static void ClearDiscordServer(@NotNull Guild guild) {
@@ -145,8 +174,11 @@ public class ExportImport {
         }
         if (!guild.getRoles().isEmpty()) {
             for (Role role : guild.getRoles()) {
-                if (guild.getSelfMember().canInteract(role)) {
-                    role.delete().queue();
+                if (guild.getSelfMember().canInteract(role) && !role.isManaged()) {
+                    try {
+                        role.delete().queue();
+                    } catch (Exception ignored) {
+                    }
                 }
             }
         }
@@ -173,55 +205,82 @@ public class ExportImport {
     }
 
     public static void ImportBackup(@NotNull JSONObject json, @NotNull Guild guild) {
-        guild.getManager().setName(json.get("name").toString()).queue();
+        guild.getManager().setName(json.getString(NAME)).queue();
+        guild.getManager().setAfkTimeout(Guild.Timeout.valueOf(json.getString(AFKTIMEOUT))).queue();
         if (guild.getFeatures().contains("COMMUNITY")) {
-            guild.getManager().setDescription(json.get("description").toString()).queue();
+            guild.getManager().setDescription(json.getString(DESCRIPTION)).queue();
         }
-        JSONObject category = json.getJSONObject("category");
+        JSONObject category = json.getJSONObject(CATEGORY);
         JSONArray jsonArray = category.names();
         for (int i = 0; i < jsonArray.length(); i++) {
-            String item = jsonArray.get(i).toString();
-            guild.createCategory(item).queue(msg -> {
-                JSONObject jsonObject = category.getJSONObject(item);
+            String item = jsonArray.getString(i);
+            JSONObject categoryJSONObject = category.getJSONObject(item);
+            int categoryPosition = categoryJSONObject.getInt(POSITION);
+            guild.createCategory(item).setPosition(categoryPosition).queue(msg -> {
+                JSONObject jsonObject = categoryJSONObject.getJSONObject(CHANNELS);
                 JSONArray channels = jsonObject.names();
-                for (int b = 0; b < channels.length(); b++) {
-                    String name = channels.get(b).toString();
-                    JSONObject channelObject = jsonObject.getJSONObject(name);
-                    String type = channelObject.getString(TYPE);
-                    if (type.equalsIgnoreCase("text")) {
-                        msg.createTextChannel(name).queue(channel -> {
-                            if (channelObject.has("Description")) {
-                                int position = channelObject.getInt(POSITION);
-                                String description = channelObject.get("Description").toString();
-                                channel.getManager().setTopic(description).queue();
-                                channel.getManager().setPosition(position).queue();
+                if (channels != null) {
+                    if (!channels.toList().isEmpty()) {
+                        for (int b = 0; b < channels.length(); b++) {
+                            String name = channels.getString(b);
+                            JSONObject channelObject = jsonObject.getJSONObject(name);
+                            String type = channelObject.getString(TYPE);
+                            if (type.equalsIgnoreCase("text")) {
+                                msg.createTextChannel(name).queue(channel -> {
+                                    if (channelObject.has(DESCRIPTION)) {
+                                        channel.getManager().setTopic(channelObject.getString(DESCRIPTION)).queue();
+                                        boolean nsfw = false;
+                                        if (channelObject.has(NSFW)) {
+                                            System.out.println("NSFW: " + channelObject.get(NSFW));
+                                            if (channelObject.getString(NSFW).equalsIgnoreCase("true")) {
+                                                nsfw = true;
+                                            }
+                                        }
+                                        System.out.println(nsfw);
+                                        channel.getManager().setNSFW(nsfw).queue();
+                                        channel.getManager().setPosition(channelObject.getInt(POSITION)).queue();
+                                    }
+                                });
+                            } else if (type.equalsIgnoreCase("voice")) {
+                                msg.createVoiceChannel(name).queue(channel -> {
+                                    channel.getManager().setPosition(channelObject.getInt(POSITION)).queue();
+                                    channel.getManager().setUserLimit(channelObject.getInt(USERLIMIT)).queue();
+                                });
+                            } else if (type.equalsIgnoreCase("stage")) {
+                                if (guild.getFeatures().contains("COMMUNITY")) {
+                                    msg.createStageChannel(name).queue(channel -> {
+                                        channel.getManager().setPosition(channelObject.getInt(POSITION)).queue();
+                                    });
+                                }
+                            } else if (type.equalsIgnoreCase("news")) {
+                                if (guild.getFeatures().contains("NEWS")) {
+                                    msg.createTextChannel(name).queue(channel -> {
+                                        channel.getManager().setType(ChannelType.NEWS).setTopic(channelObject.getString(TOPIC)).queue();
+                                        channel.getManager().setPosition(channelObject.getInt(POSITION)).queue();
+                                        channel.getManager().setNSFW(channelObject.getBoolean(NSFW)).queue();
+                                    });
+                                }
                             }
-                        });
-                    } else if (type.equalsIgnoreCase("voice")) {
-                        msg.createVoiceChannel(name).queue(channel -> {
-                            int position = channelObject.getInt(POSITION);
-                            int userLimit = channelObject.getInt(USERLIMIT);
-                            channel.getManager().setPosition(position).queue();
-                            channel.getManager().setUserLimit(userLimit).queue();
-                        });
-                    } else if (type.equalsIgnoreCase("stage")) {
-                        if (guild.getFeatures().contains("COMMUNITY")) {
-                            msg.createStageChannel(name).queue(channel -> {
-                                int position = channelObject.getInt(POSITION);
-                                channel.getManager().setPosition(position).queue();
-                            });
-                        }
-                    } else if (type.equalsIgnoreCase("news")) {
-                        if (guild.getFeatures().contains("NEWS")) {
-                            msg.createTextChannel(name).queue(m -> {
-                                m.getManager().setType(ChannelType.NEWS).queue();
-                                int position = channelObject.getInt(POSITION);
-                                m.getManager().setPosition(position).queue();
-                            });
                         }
                     }
                 }
             });
+        }
+
+        if (json.has(AFKCHANNEL)) {
+            List<VoiceChannel> afkchannel = guild.getVoiceChannelsByName(json.getString(AFKCHANNEL), true);
+            if (!afkchannel.isEmpty()) {
+                guild.getManager().setAfkChannel(afkchannel.get(0)).queue();
+            }
+        }
+        if (json.has(DEFAULTNOTIFICATIONLEVEL)) {
+            guild.getManager().setDefaultNotificationLevel(Guild.NotificationLevel.valueOf(json.getString(DEFAULTNOTIFICATIONLEVEL))).queue();
+        }
+        if (json.has(SYSTEMMESSAGECHANNEL)) {
+            List<TextChannel> systemchannel = guild.getTextChannelsByName(json.getString(SYSTEMMESSAGECHANNEL), true);
+            if (!systemchannel.isEmpty()) {
+                guild.getManager().setSystemChannel(systemchannel.get(0)).queue();
+            }
         }
 
         JSONObject roles = json.getJSONObject(ROLES);
